Linux/ELF: Add option to scan loaded shared libraries
Dynamic linking is the standard on Linux.  Thus the hardening of dynamic
loaded shared libraries can affect programs as well.

Add a an option to scan all dynamic libraries for a binary/process.  For
binaries the list of libraries is taken from the ELF information and for
processes all executable mapped memory regions backed up by a file are
scanned.